Bug fixes

* security fix and bugfix

* user registration bugfix

* user registration bugfix

* fix category in add auction

* fixes

Approved-by: Imre Boersma <imre.boersma@gmail.com>

+++ src/helpers/GeneralHelpers.php
@@ -34,5 +34,5 @@ function active_url($url, $callback = '')
 
 function numberToEuro($number)
 {
-    return "€ " . number_format(intval($number), 2, ',', ' ');
+    return "€ " . number_format(floatval($number), 2, ',', ' ');
 }
